fix: [Refactoring] organized the projects domain folder
