+++ src/domain/projects/ProjectsData.ts
@@ -1,11 +1,12 @@
-import { ProjectType } from "../../utils/MultiCardsIntetrfaces";
+import { ProjectType } from '../../utils/MultiCardsIntetrfaces';
 
 export const ProjectsData: ProjectType[] = [
   {
-    title: "Filler Project",
+    title: 'Filler Project',
     description:
-      "Filler project that work as a placeholder for the rest of the projects. It helps to show the use case for the pagination and filters.",
-    img_path: "/imgs/projects/thumbnails/filler.png",
-    tags: ["Filler", "Project", "NoData"],
+      'Filler project that work as a placeholder for the rest of the projects. It helps to show the use case for the pagination and filters.',
+    img_path: '/imgs/projects/thumbnails/filler.png',
+    tags: ['Filler', 'Project', 'NoData'],
+    redirect_url: '',
   },
 ];
