+++ src/domain/projects/components/filler/FillerProject.tsx
@@ -1,4 +1,4 @@
-import WorkInProgress from '../../WorkingInProgress';
+import WorkInProgress from '../WorkingInProgress';
 
 const FillerProject: React.FC = function fillerProject() {
   return <WorkInProgress project="Filler Project" />;
