+++ src/domain/projects/components/nft/Nft.tsx
@@ -1,4 +1,4 @@
-import WorkInProgress from "../WorkingInProgress";
+import WorkInProgress from '../WorkingInProgress';
 
 const Nft: React.FC = function nft() {
   return <WorkInProgress project="NFT Shop" />;
