+++ src/domain/projects/components/podcast/PodOfCast.tsx
@@ -1,4 +1,4 @@
-import WorkInProgress from "../WorkingInProgress";
+import WorkInProgress from '../WorkingInProgress';
 
 const PodOfCast: React.FC = function podOfCast() {
   return <WorkInProgress project="Pod Of Cast" />;
