+++ src/domain/projects/components/youlearn/YouLearn.tsx
@@ -1,4 +1,4 @@
-import WorkInProgress from "../WorkingInProgress";
+import WorkInProgress from '../WorkingInProgress';
 
 const YouLearn: React.FC = function youLearn() {
   return <WorkInProgress project="YouLearn Online" />;
