+++ src/domain/projects/components/yoursmile/YourSmile.tsx
@@ -1,4 +1,4 @@
-import WorkInProgress from "../WorkingInProgress";
+import WorkInProgress from '../WorkingInProgress';
 
 const YourSmile: React.FC = function yourSmile() {
   return <WorkInProgress project="Your Smile" />;
